implemented error handaling

+++ src/main/java/com/in28minutes/springboot/web/springbootfirstwebapplication/TodoController.java
@@ -53,6 +53,10 @@ public class TodoController {
 	
 	@RequestMapping(value="/delete-todo", method = RequestMethod.GET)
 	public String deleteTodo(@RequestParam int id) {
+		
+		if(id==1)
+			throw new RuntimeException("Something went wrong");
+		
 		todoService.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
